+++ packages/api/src/business/validations/document-updation-validate.ts
@@ -1,6 +1,5 @@
+import { dateValidate } from 'camino-common/src/date.js'
 import { IDocument } from '../../types.js'
-
-import { dateValidate } from '../../tools/date.js'
 
 export const documentUpdationValidate = async (document: IDocument) => {
   const errors = [] as string[]
